CursosAnthropC/IntroductionToModelContextProtocol/project_mcp_v1/requirements.txt
fastapi==0.135.2
typer==0.24.1
rich==14.3.3
openai==2.30.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
loguru==0.7.3
httpx==0.28.1
mcp==1.26.0
anyio==4.13.0
uvicorn[standard]==0.42.0
aiomysql==0.2.0
